Move subscription tracking logic over to batchActions file
